Email addresses not collected, bug list limited to 500 bugs
Since a long time ago, the API (at R bug tracker) does not return email
addresses in some places where it used to. Solution: do not try to collect
that data.

A new argument 'limit' was added to list_bugs(), with default 500.
This limits the number of bugs returned by the function. The reason for
this change is some problems (502 bad gateway) seen when trying to fetch
a large or unlimited number of bugs.

Documents were updated (with roxygen2).

Version and Date were updated in DESCRIPTION.